refactor: generation 필드 Integer로 변경 (#283)

+++ src/main/java/org/sopt/app/application/auth/PlaygroundAuthService.java
@@ -169,7 +169,7 @@ public class PlaygroundAuthService {
         return playgroundClient.getOwnPlaygroundProfile(requestHeader);
     }
 
-    public List<PlaygroundAuthInfo.PlaygroundProfileOfRecommendedFriend> getPlaygroundProfilesForSameGeneration(Long generation) {
+    public List<PlaygroundAuthInfo.PlaygroundProfileOfRecommendedFriend> getPlaygroundProfilesForSameGeneration(Integer generation) {
         return playgroundClient.getPlaygroundProfileForSameGeneration(createAuthorizationHeader(playgroundToken), generation);
     }
 }
